Do not exit rcubic when last step fails
A bug was introduced with exectree migration, rcubic used to wait for
all script to succeed (unless in session mode). However the bug changed
behaviour, where rcubic could exit with unfinished (but failed) steps
-- so long as they were the last ones.

Add a waitsuccess mode to exectree, which when set does not complete
blocking run untill all steps are sucessfully. If waitsuccess is set to
False (default), then it will exit as soon as all jobs have finished
execution regardless of execution status.

Change-Id: Iab170dbd5589492baf0b721d113c9411b7f77abc